requirements.txt
asgiref==3.7.2
bcrypt==4.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
django-cors-headers==4.2.0
django-ninja==0.22.2
gunicorn==21.2.0
idna==3.4
packaging==23.1
Pillow==10.0.0
psycopg2-binary==2.9.6
pydantic==1.10.12
PyJWT==2.8.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.4
